requirements.txt
Flask==2.2.3
gunicorn==20.1.0
template-render==0.1.0
opencv-python==4.7.0.68
numpy==1.24.2
Pillow==9.4.0
pytesseract==0.3.10
//keras==2.11.0
//streamlit==1.18.1